SPPy/Langchain CodeBasics/requirements.txt
pinecone-client
ipykernel                       # for ipynb and interactive partial  kernel execution
python-dotenv                   # MANDATORY to load env viaraoble from .env
numpy
numexpr

#User GUI
streamlit                       # quick web by streamlit run main.py
panel                           #The Powerful Data Exploration & Web App Framework for Python
streamlit_lottie                # why? from ai_interviewer
jupyter_bokeh                   #Using Panel interactively in VSCode notebooks requires the jupyter_bokeh package to be installed

langchain                       #Deprectade 
langchain_community
langchain_openai
#langchain_agents                # for Tools. from langchain.agents import tool
                                # had to nad upgarde pip in virt env 'curl https://bootstrap.pypa.io/get-pip.py | python'
                                # AND then faile dto install langchain_agents again
                                # S0 additionally  had to manually upgrade  do 'pip install --upgrade langchain'
# got error installing langchain_agents 
# 'Could not find a version that satisfies the requirement langchain_agents (from versions: none)'
# https://stackoverflow.com/questions/49745105/could-not-find-a-version-that-satisfies-the-requirement-in-python
# above asks to 'pip install --upgrade virtenv' did not wwork
# https://stackoverflow.com/questions/49748063/pip-install-fails-for-every-package-could-not-find-a-version-that-satisfies/49748494#49748494
# above asks to 'curl https://bootstrap.pypa.io/get-pip.py | python' in virtenv
# AND then faile dto install langchain_agents again
# S0 additionally  had to manually upgrade  do 'pip install --upgrade langchain'

#not required seprately pat of lib???? langchain_core                  # for pydantic function calling langchain_core.utils.function_calling.convert_to_openai_functio
openai
huggingface_hub
wikipedia
groq
#not required seprately pat of lib???? datetime                        # datetime.now().strftime("%d%b%y %H:%M:%S")

#huggingface_hubpip 

boto3                            # for AWS Bedrock
sagemaker
pinecone-client
ipywidgets
boto3_helpers                   # For logging on cloudwatch import yield_metric_data


#tools
#google-search-results          #serpapi
duckduckgo-search              # Free comapred to serpapi
#numexpr #llmmathchain
#langgraph                      # >2 agent pipeline
#langchain_openai
#langchainhub
#langsmith                      # UI to see langchain status
#beautifulsoup4
#gradio


#tensorflow crashcourse
#setuptools
#deprecated sklearn
#scikit-learn
#tensorflow
#python -m pip show tensorflow --- shoudl show tensorflow
#keras


openapi-pydantic            # to get API's working  -- https://github.com/langchain-ai/langchain/issues/9520
                            # To fix AttributeError: 'super' object has no attribute 'parse_obj' using OpenAPISpec.from_text(text) 
PyPDF2